+++ client/src/App.css
@@ -33,10 +33,6 @@
   background-image: var(--gradient-primary);
 }
 
-/* .login-signup-form {
-  position: relative;
-} */
-
 .login-signup-form {
   background-image: var(--gradient-primary);
   border-radius: 25px;
@@ -164,4 +160,70 @@
 
 .chart {
   margin: 0 auto;
+}
+
+.btn {
+  background: radial-gradient(
+    circle at -1% 57.5%,
+    rgb(19, 170, 82) 0%,
+    rgb(0, 102, 43) 90%
+  );
+  color: #fff;
+  border: none;
+  position: relative;
+  height: 60px;
+  font-size: 1.6em;
+  padding: 0 2em;
+  cursor: pointer;
+  transition: 800ms ease all;
+  outline: none;
+}
+.btn:hover {
+  background: #fff;
+  color: #1aab8a;
+}
+.btn:before,
+.btn:after {
+  content: "";
+  position: absolute;
+  top: 0;
+  right: 0;
+  height: 2px;
+  width: 0;
+  background: #1aab8a;
+  transition: 400ms ease all;
+}
+.btn:after {
+  right: inherit;
+  top: inherit;
+  left: 0;
+  bottom: 0;
+}
+.btn:hover:before,
+.btn:hover:after {
+  width: 100%;
+  transition: 800ms ease all;
+}
+
+.signup-login-href {
+  position: relative;
+}
+
+.signup-login-href::before {
+  content: "";
+  position: absolute;
+  width: 100%;
+  height: 4px;
+  border-radius: 4px;
+  background-color: #18272f;
+  bottom: 0;
+  left: 0;
+  transform-origin: right;
+  transform: scaleX(0);
+  transition: transform 0.3s ease-in-out;
+}
+
+.signup-login-href:hover::before {
+  transform-origin: left;
+  transform: scaleX(1);
 }
